+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed: Improved server performance when importing/generating games with a huge number of worlds.
 
+### Metroid: Samus Returns
+
+- Added: Option to change the final boss to be Arachnus, Diggernaut, Metroid Queen, or Proteus Ridley (default).
+
 ## [8.4.0] - 2024-09-04
 
 - Fixed: The "Last Activity" text on the Multiplayer Session window is not aligned properly.
@@ -53,7 +57,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Metroid: Samus Returns
 
-- Added: Option to change the final boss to be Arachnus, Diggernaut, Proteus Ridley (default), or the Metroid Queen.
 - Added: Changing the volume of the music and background ambience is now possible via cosmetic options.
 - Added: Option to have more Metroid DNA placed than required.
 - Changed: Area 6 - The Door to Chozo Seal West Intersection Terminal from Crumbling Stairwell is excluded from Door Lock Rando.